+++ src/main/java/com/gaebang/backend/domain/ai/service/RecommendationService.java
@@ -1,11 +1,11 @@
-package com.gaebang.backend.domain.aianalysis.service;
+package com.gaebang.backend.domain.ai.service;
 
-import com.gaebang.backend.domain.aianalysis.entity.AIModelIntegrated;
-import com.gaebang.backend.domain.aianalysis.recommendation.Answers;
-import com.gaebang.backend.domain.aianalysis.recommendation.Model;
-import com.gaebang.backend.domain.aianalysis.recommendation.ModelScoreCalculator;
-import com.gaebang.backend.domain.aianalysis.recommendation.Weights;
-import com.gaebang.backend.domain.aianalysis.repository.AIAnalysisRepository;
+import com.gaebang.backend.domain.ai.entity.AIModelIntegrated;
+import com.gaebang.backend.domain.ai.recommendation.Answers;
+import com.gaebang.backend.domain.ai.recommendation.Model;
+import com.gaebang.backend.domain.ai.recommendation.ModelScoreCalculator;
+import com.gaebang.backend.domain.ai.recommendation.Weights;
+import com.gaebang.backend.domain.ai.repository.AIAnalysisRepository;
 import com.gaebang.backend.global.util.ResponseDTO;
 import org.springframework.stereotype.Service;
